+++ pkg/controllers/node/controller.go
@@ -50,9 +50,8 @@ func NewController(clk clock.Clock, kubeClient client.Client, cloudProvider clou
 	return &Controller{
 		kubeClient:     kubeClient,
 		cluster:        cluster,
-		initialization: &Initialization{kubeClient: kubeClient, cloudProvider: cloudProvider},
 		emptiness:      &Emptiness{kubeClient: kubeClient, clock: clk, cluster: cluster},
-		expiration:     &Expiration{kubeClient: kubeClient, clock: clk},
+		initialization: &Initialization{kubeClient: kubeClient, cloudProvider: cloudProvider},
 	}
 }
 
@@ -62,9 +61,8 @@ type Controller struct {
 	kubeClient     client.Client
 	cluster        *state.Cluster
 	initialization *Initialization
-	emptiness      *Emptiness
-	expiration     *Expiration
 	finalizer      *Finalizer
+	emptiness      *Emptiness
 }
 
 // Reconcile executes a reallocation control loop for the resource
@@ -96,7 +94,6 @@ func (c *Controller) Reconcile(ctx context.Context, req reconcile.Request) (reco
 		Reconcile(context.Context, *v1alpha5.Provisioner, *v1.Node) (reconcile.Result, error)
 	}{
 		c.initialization,
-		c.expiration,
 		c.emptiness,
 		c.finalizer,
 	} {
